refactor: fix more clang-tidy warnings

+++ plugin/src/Caelestia/Services/audiocollector.cpp
@@ -202,7 +202,7 @@ void AudioCollector::loadChunk(const qint16* samples, quint32 count) {
 
     auto* writeBuffer = m_writeBuffer.load(std::memory_order_relaxed);
     std::transform(samples, samples + count, writeBuffer->begin(), [](qint16 sample) {
-        return sample / 32768.0f;
+        return static_cast<float>(sample) / 32768.0f;
     });
 
     auto* oldRead = m_readBuffer.exchange(writeBuffer, std::memory_order_acq_rel);
